requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
pandas==0.24.2
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
roboatenviro==0.1.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
